requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-filter==23.2
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
mailgun==0.1.1
psycopg2-binary==2.9.7
pyactiveresource==1.0.1
PyJWT==2.8.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
secure-smtplib==0.1.1
sqlparse==0.4.4
stripe==6.3.0
typing_extensions==4.7.1
urllib3==2.0.4
